requirements.txt
Flask==2.2.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
itsdangerous==2.0.1
Pillow==9.5.0
WTForms==3.0.1
gunicorn==20.1.0
email_validator==1.3.1
Werkzeug==2.2.2
wtforms==2.2.1